I think step 2b is done

+++ lab6/list.c
@@ -230,7 +230,7 @@ int Remove_From_End(linked_list_t list, int* data)
 // Params:
 //    list: list to traverse
 //    action: function to call with each data item
-int Traverse(linked_list_t list, void (*action)(int64_t data))
+int Traverse(linked_list_t list, void (*action)(int data))
 {
 	int err = 0;
 	if(list==NULL || action==NULL)
